+++ src/init.spec.lua
@@ -13,6 +13,9 @@ return function()
 			update = "function",
 			oneChild = "function",
 			setGlobalConfig = "function",
+			createContext = "function",
+			provide = "function",
+			consume = "function",
 
 			-- These functions are deprecated and throw warnings!
 			reify = "function",